Mark TaskThreadPool::sMutexRW as hidden
The same symbol will be present in loadable modules, too. When the
loadable modules are loaded, the constructors for sMutexRW within
the module will be run, but sMutexRW in the modules will point to
the one in the main executable instead of their own, if the symbol
is globally visible. This leads double construction and destruction,
which causes hangs on exit on x86_64.

From dss-hh-20081021-1.patch by Horace Hsieh