Make sure we can send RTP-Info headers if the reflection buffer size is zero
Store packets for 2 seconds if the duration otherwise would be zero. Look
for packets for RTP-Info at 150 ms behind the buffering duration if the
RTP info packet offset is higher than the buffering duration.

This makes sure the server is able to send proper RTP-Info headers even
when doing zero-latency streaming.