Don't use global ip component string objects for the local addresses
Instead just declare them locally in IsLocal. This requires
redoing the parsing of them for each IsLocal call.

This avoids having them as global data both in the main server
and in the dynamically loaded modules. The constructors/destructors
for the global object causes issues when there are multiple
definitions of the same global object.

This fixes crashes in some setups, crashes that related to
running of the constructors/destructors of these global objects,
when the dynamic modules were involved.