Clarify and clean up reference counting in the reflector module
Previously, the whole ReflectorSession and associated UDP sockets
could easily be leaked in various (not too uncommon) error cases.
One example was if a client tries to describe and play back a live
broadcaster session after it has been torn down, while some other
clients still have references to it (so the DoDescribe call first
got a reference to the ReflectorSession, but later failed due to
the SDP file missing).

Making it leak-free seems to require slightly changing the scheme
of how the reference counts are supposed to be handled. Now, the
reference counting scheme is as follows:

No permanent reference count should be added by DESCRIBE, nor by
RECORD (since we can't undo that one properly in DestroySession).
Only add one reference count by SETUP, regardless of the number of
streams (since SETUP is the method that creates the ReflectorSession
object or the ReflectorOutput objects, and we can't easily find the
actual number of set up streams in RemoveOutput), and remove
reference counts similiarly in the RemoveOutput function.